Bugfixes, cleanup, log tags

+++ Engine/Source/Graphics/Render/Render.h
@@ -34,7 +34,7 @@ void RenderSmokeGlobal(int64_t iCommandBuffer);
 
 // Wind
 inline bool gbWindClear = false;
-inline int giWindTextureIndex = 0; // 0 = write TextureOne, 1 = write TextureTwo
+inline int64_t giWindTextureIndex = 0; // 0 = write TextureOne, 1 = write TextureTwo
 
 void RenderWindGlobal(int64_t iCommandBuffer);
